chore: Add typescript definitions
